🔧 Chore : 에러 표시 아이콘 컴포넌트 분할
