+++ 자이-백병재/3주차/mission-2/src/components/ErrorAppear.tsx
@@ -1,3 +1,5 @@
+import ErrorIcon from "./ErrorIcon";
+
 const ErrorAppear = () => {
   return (
     <div className="flex min-h-screen flex-col items-center justify-center bg-white text-red-700 p-6">
@@ -10,18 +12,6 @@ const ErrorAppear = () => {
     </div>
   );
 };
-
-const ErrorIcon: React.FC = () => (
-  <svg 
-    xmlns="http://www.w3.org/2000/svg" 
-    className="h-24 w-24 text-red-700"
-    fill="none" 
-    viewBox="0 0 24 24" 
-    stroke="currentColor" 
-  >
-    <path strokeLinecap="round" strokeLinejoin="round" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-  </svg>
-);
 
 
 export default ErrorAppear;
