+++ 자이-백병재/3주차/mission-2/src/components/ErrorIcon.tsx
@@ -1,0 +1,13 @@
+const ErrorIcon: React.FC = () => (
+  <svg 
+    xmlns="http://www.w3.org/2000/svg" 
+    className="h-24 w-24 text-red-700"
+    fill="none" 
+    viewBox="0 0 24 24" 
+    stroke="currentColor" 
+  >
+    <path strokeLinecap="round" strokeLinejoin="round" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+  </svg>
+);
+
+export default ErrorIcon
